# apps/vuln-api/app/blueprints/security_ops/routes.py
"""
Routes for security ops endpoints.
"""

from flask import request, jsonify, render_template_string, session
from datetime import datetime, timedelta
import uuid
import random
import json
import time

from . import security_ops_bp
from app.models import *

@security_ops_bp.route('/api/security/monitoring/bypass')
def security_monitoring_bypass():
    """Bypass security monitoring"""
    return jsonify({
        'strategies': ['log_suppression', 'traffic_obfuscation', 'alert_saturation'],
        'monitoring_disabled': True,
        'duration_minutes': 45
    })


@security_ops_bp.route('/api/incidents/create', methods=['POST'])
def incidents_create():
    """Create security incident record"""
    data = request.get_json() or {}
    incident_type = data.get('incident_type', '')
    severity = data.get('severity', 'medium')
    description = data.get('description', '')
    affected_systems = data.get('affected_systems', [])


@security_ops_bp.route('/api/threats/indicators')
def threats_indicators():
    """Get threat intelligence indicators"""
    return jsonify({
        'indicators': [
            {
                'type': 'ip_address',
                'value': '192.0.2.100',
                'threat_level': 'high',
                'category': 'C2-Server',
                'first_seen': '2024-12-01T10:00:00Z',
                'last_seen': '2024-12-08T14:30:00Z',
                'confidence': 95,
                'source': 'ThreatIntel-Feed-A'
            },
            {
                'type': 'domain',
                'value': 'malicious-domain.xyz',
                'threat_level': 'critical',
                'category': 'Phishing',
                'first_seen': '2024-12-05T08:00:00Z',
                'last_seen': '2024-12-08T15:00:00Z',
                'confidence': 98,
                'source': 'OSINT'
            },
            {
                'type': 'file_hash',
                'value': 'a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6',
                'threat_level': 'high',
                'category': 'Ransomware',
                'first_seen': '2024-12-07T12:00:00Z',
                'last_seen': '2024-12-08T13:45:00Z',
                'confidence': 92,
                'source': 'Sandbox-Analysis'
            },
            {
                'type': 'url',
                'value': 'http://malware-download.bad/payload.exe',
                'threat_level': 'critical',
                'category': 'Malware-Distribution',
                'first_seen': '2024-12-06T15:00:00Z',
                'last_seen': '2024-12-08T16:00:00Z',
                'confidence': 97,
                'source': 'Web-Crawler'
            }
        ],
        'total_indicators': 4,
        'indicators_blocked': 4,
        'threat_feeds_active': 12,
        'last_update': datetime.now().isoformat(),
        'ioc_format': 'STIX 2.1'
    })


@security_ops_bp.route('/api/remediation/apply', methods=['POST'])
def remediation_apply():
    """Apply security remediation actions"""
    data = request.get_json() or {}
    remediation_type = data.get('remediation_type', '')
    target_systems = data.get('target_systems', [])
    automated = data.get('automated', True)


@security_ops_bp.route('/api/security/posture/harden', methods=['PUT'])
def security_posture_harden():
    """Harden security posture"""
    data = request.get_json() or {}
    hardening_profile = data.get('hardening_profile', 'standard')
    scope = data.get('scope', 'all')


@security_ops_bp.route('/api/patches/deploy', methods=['POST'])
def patches_deploy():
    """Deploy security patches"""
    data = request.get_json() or {}
    patch_ids = data.get('patch_ids', [])
    target_systems = data.get('target_systems', [])
    deployment_window = data.get('deployment_window', 'maintenance')


@security_ops_bp.route('/api/security/alerts/acknowledge', methods=['POST'])
def security_alerts_acknowledge():
    """Acknowledge security alerts"""
    data = request.get_json() or {}
    alert_ids = data.get('alert_ids', [])
    analyst_id = data.get('analyst_id', '')
    notes = data.get('notes', '')


@security_ops_bp.route('/api/defense/metrics')
def defense_metrics():
    """Get defensive security metrics"""
    return jsonify({
        'metrics_period': '24_hours',
        'detection_metrics': {
            'total_events_processed': random.randint(1000000, 5000000),
            'alerts_generated': random.randint(500, 2000),
            'true_positives': random.randint(50, 200),
            'false_positives': random.randint(100, 500),
            'detection_rate': random.uniform(85, 98),
            'false_positive_rate': random.uniform(5, 15)
        },
        'response_metrics': {
            'incidents_created': random.randint(10, 50),
            'incidents_resolved': random.randint(8, 45),
            'mean_time_to_detect_minutes': random.randint(5, 30),
            'mean_time_to_respond_minutes': random.randint(15, 60),
            'mean_time_to_resolve_hours': random.uniform(2, 12),
            'automation_rate': random.uniform(60, 85)
        },
        'threat_metrics': {
            'threats_blocked': random.randint(1000, 5000),
            'malware_detected': random.randint(50, 200),
            'phishing_attempts': random.randint(100, 500),
            'brute_force_attacks': random.randint(200, 800),
            'ddos_attempts': random.randint(5, 20)
        },
        'vulnerability_metrics': {
            'vulnerabilities_discovered': random.randint(10, 50),
            'vulnerabilities_remediated': random.randint(15, 45),
            'mean_time_to_patch_days': random.uniform(3, 15),
            'systems_scanned': random.randint(500, 2000),
            'patch_compliance_rate': random.uniform(85, 98)
        },
        'timestamp': datetime.now().isoformat(),
        'dashboard_url': 'https://soc.example.com/metrics'
    })