# apps/vuln-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title:  TestBed API (api-demo)
  version: 1.0.0
  description: >-
    Static OpenAPI spec for the api-demo app. For full narrative docs see
    API-DOCUMENTATION.md. Endpoints are intentionally vulnerable for testing.
servers:
  - url: http://localhost:8880
...
paths:
...
  /api/incidents/create:
    post:
      tags: [Security Ops]
      summary: Post
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
...
  /api/patches/deploy:
    post:
      tags: [Security Ops]
      summary: Post
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
...
  /api/remediation/apply:
    post:
      tags: [Security Ops]
      summary: Post
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
...
  /api/security/alerts/acknowledge:
    post:
      tags: [Security Ops]
      summary: Post
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
...
  /api/security/posture/harden:
    put:
      tags: [Security Ops]
      summary: Put
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
...